test/strategy/replaceconditonals/IfElseDemoTest.java
package strategy.replaceconditonals;

import org.junit.Test;

import static org.junit.Assert.*;

public class IfElseDemoTest {
    @Test
    public void low() {
        assertEquals(1825, insuranceFor(5000), 0.01);
    }

    @Test
    public void medium() {
        assertEquals(38600, insuranceFor(25000), 0.01);
    }

    @Test
    public void high() {
        assertEquals(78500, insuranceFor(50000), 0.01);
    }

    @Test
    public void veryHigh() {
        assertEquals(106400, insuranceFor(100000), 0.01);
    }

    private double insuranceFor(double income) {
        return new IfElseDemo().calculateInsurance(income);
    }
}